requirements.txt
numpy==2.1.1
scipy==1.13.1
pandas==2.2.2
matplotlib==3.8.4
pyyaml==6.0.2
jinja2==3.1.4
statsmodels==0.14.2
fastapi==0.115.0
uvicorn==0.30.6
scikit-learn==1.5.2
